pyproject.toml
[tool.poetry]
name = "pyeuropeana"
version = "0.1.7"
description = "A Python wrapper around Europeana APIs"
authors = [
    "José Eduardo Cejudo Grano de Oro <joseed.cejudo@europeana.eu>",
    "Eren Janberk Genç <erenjanberk.genc@europeana.eu>",
]
license = "European Union Public Licence v. 1.2"
readme = "README.md"
homepage = "https://github.com/europeana/rd-europeana-python-api"
repository = "https://github.com/europeana/rd-europeana-python-api"
keywords = ["GLAM", "europeana", "cultural heritage"]
classifiers = [
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research"
]
packages = [
    { include = "pyeuropeana", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.7.1, <4.0"
requests = "^2.27"
pandas = "^1.3"
pillow = "7.1.2"
fire = "^0.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2"
pytest = "^7"
nbconvert = "^6.5"
sphinx = "^4.5"
sphinx-rtd-theme = "1.0.0"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
